- Completed startup and reset functionalities - Documentation

+++ src/main/java/com/sashacorp/springmongojwtapi/appconfig/AppConfiguratorController.java
@@ -10,14 +10,17 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.sashacorp.springmongojwtapi.models.persistence.user.Authority;
 import com.sashacorp.springmongojwtapi.models.persistence.user.User;
-import com.sashacorp.springmongojwtapi.repository.LeaveRepository;
+import com.sashacorp.springmongojwtapi.repository.EventTypeRepository;
 import com.sashacorp.springmongojwtapi.repository.MessageRepository;
 import com.sashacorp.springmongojwtapi.repository.UserRepository;
+import com.sashacorp.springmongojwtapi.security.Authority;
 import com.sashacorp.springmongojwtapi.util.http.HttpUtil;
+import com.sashacorp.springmongojwtapi.util.http.Par;
+import com.sashacorp.springmongojwtapi.util.http.Url;
 import com.sashacorp.springmongojwtapi.util.http.hateoas.HateoasUtil;
 
 
@@ -34,16 +37,16 @@ public class AppConfiguratorController {
 	@Autowired
 	MessageRepository messageRepository;
 	@Autowired
-	LeaveRepository leaveRepository;
+	EventTypeRepository eventTypeRepository;
 	
 	
 	/**
 	 * Endpoint for registering super-administrator on the first run.
-	 * After registration the mapping is switched off.
+	 * After registration the '/startup' mapping is switched off.
 	 * @param startupRequest
 	 * @return
 	 */
-	@PostMapping("/startup")
+	@PostMapping(Url.STARTUP)
 	public ResponseEntity<?> registerFirstUser(@RequestBody StartupRequest startupRequest) {
 		if (appConfigurator.getAppConfiguration() == null) {
 			return HttpUtil.getHttpStatusResponse(HttpStatus.NOT_FOUND);
@@ -57,7 +60,7 @@ public class AppConfiguratorController {
 				startupRequest.getUsername());
 		
 		AppConfiguration appConfig = new AppConfiguration(startupRequest.getCompanyName(), true);
-
+		
 		Set<Authority> authorities = new HashSet<>();
 		
 		authorities.add(Authority.ADMIN);
@@ -74,22 +77,30 @@ public class AppConfiguratorController {
 	
 	/**
 	 * Endpoint for a complete clean-up of all resources and configuration.
+	 * If the resetKey parameter is not present, returns the current application configuration UUID.
+	 * If the resetKey parameter is present, accepts the current application configuration UUID and resets all collections and state.
 	 * Admin only.
 	 * @param startupRequest
 	 * @return
 	 */
-	@PostMapping("/reset")
-	public ResponseEntity<?> stopAndClean(@RequestBody StartupRequest startupRequest) {
+	@PostMapping(Url.RESET)
+	public ResponseEntity<?> stopAndClean(@RequestParam(name = Par.RESET_KEY, required = false) String resetKey) {
 		if (appConfigurator.getAppConfiguration() == null) {
 			return HttpUtil.getHttpStatusResponse(HttpStatus.NOT_FOUND);
 		}
-		
-		userRepository.deleteAll();
-		messageRepository.deleteAll();
-		leaveRepository.deleteAll();
-		AppConfiguration appConfig = new AppConfiguration("", false);	
-		AppConfiguration savedConfig = appConfigurator.setAppConfiguration(appConfig);
-		return HttpUtil.getResponse(savedConfig, HttpStatus.CREATED);
+		if (resetKey == null) {
+			resetKey = appConfigurator.getAppConfiguration().getId();
+			return HttpUtil.getPlainTextResponse(resetKey, HttpStatus.OK);
+		} else if (resetKey.equals(appConfigurator.getAppConfiguration().getId())) {
+			userRepository.deleteAll();
+			messageRepository.deleteAll();
+			eventTypeRepository.deleteAll();
+			AppConfiguration appConfig = new AppConfiguration("", false);	
+			AppConfiguration savedConfig = appConfigurator.setAppConfiguration(appConfig);
+			return HttpUtil.getResponse(savedConfig, HttpStatus.CREATED);			
+		} else {
+			return HttpUtil.getHttpStatusResponse(HttpStatus.FORBIDDEN);
+		}
 	}
 	
 }
